Added functionality to year, month, week buttons in chart
I added functionality to the year, month, and week buttons by creating a click handler method and a data retrieval method. The click handler method decided if the data retrieval method should be called. If yes, then the new data is retrieved and the state is set to the new data.